abstracted out articles routes
